XER10-2417 : [SKY-VL-QE][Sprint][Migration]:Wanamanger is continously restarting during the downgrade migration
Reason for change:
As part of RDKB-59763 US, we have supported SLAAC IPv6 support in WanManager but that new flag has not been handled as part of t2 profile due to the reason we seen crash. so to avoid this crash I have handled this in t2 profile parsing during t2 event send.
Handled Boundary Check Mechanism for some of the Virtual Interface Modes to avoid segemenation fault.

Test Procedure:
1. WANManager should not crash and should work without any issue.
2. Upgrade and Downgrade case should not cause any issue.

Risks: Medium

Signed-off-by: LakshminarayananShenbagaraj <lakshminarayanan.shenbagaraj2@sky.uk>